requirements.txt
numpy==1.22.3
scipy==1.8.0
matplotlib==3.5.2
pandas==1.4.2
jupyter==1.0.0
opencv-python==4.5.5.64
Pillow==9.1.0
Flask==2.1.2
gunicore==20.1.0